python extension reorg
