+++ python/pdal/__init__.py
@@ -1,1 +1,3 @@
 __version__='1.3.0'
+
+from .pipeline import Pipeline
